End to End ML Project/twitter_gpt3_project/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
numpy==1.22.2
oauthlib==3.2.0
pandas==1.4.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
tweepy==4.6.0
twitter-application-only-auth==0.3.3
urllib3==1.26.8
